pyproject.toml
[tool.poetry]
name = "highlights"
version = "0.1.0"
description = ""
authors = ["Simon Bernier St-Pierre <sbernierstpierre@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
nhlapi = {git = "https://github.com/reddit-habs/nhlapi.git"}
requests = "^2.21"
attrs = "^18.2"
jinja2 = "^2.10"
pendulum = "^2.0"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
